user-parking-control-service/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-RESTful==0.3.8
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
flask-mysqldb
PyJWT==2.0.0
PyMySQL==0.10.1
flask-cors
datetime
uuid
